backend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
inflection==0.5.1
packaging==24.2
platformdirs==4.3.6
psycopg2==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.28.0
